+++ CHANGELOG.md
@@ -34,6 +34,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - GmshParser: define_quadratic_tetrahedron_element_mesh
 - GmshParser: to_vtk (consider quadratic tetrahedron element)
 - Assembly: define a volmdlr Assembly object.
+- Edge: direction_independent_is_close
+- Arcellipse2D, 3D: complementary
+- 
 
 ### Fixed
 - 2D conversion: create 2D function name in core_compiled
@@ -84,6 +87,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Arc3D.minimum_distance_points_line
 - New unittests for plane3d
 - ClosedShel3D: intersection
+- Arcellipse2D: complementary
 
 ## v0.10.0 [Unreleased yet]